requirements.txt
streamlit==0.73.0
tensorflow-cpu==2.4.0
numpy==1.19.4
nltk==3.5
#keras==2.3.1


